layers: Audited renderpass validation, adding many checks alongside CRP2 support
To properly validate VK_KHR_create_renderpass2 (CRP2), the first step was to ensure all the validation for the v1 versions of each new function was up to scratch, as rather than effectively duplicating all the validation those calls have already, it made much more sense to add a switch and reuse the existing code where reasonable.
There dozens of missing VUID checks (and a couple of mislabeld ones) compared to what is in the spec for the existing vkCreateRenderPass, vkCmdBeginRenderPass, vkCmdNextSubpass, vkCmdEndRenderPass, and vkDestroyRenderPass functions (and their sub structs) to match up with what's already there.
Almost all of these VUID checks have now been implemented - notable omissions are allocator callback validation (not sure how to tackle that, and needs to be done object-wide) and some DeviceGroup functionality.
Some VUIDs that were not actually validatable or were duplicates were skipped as well, and these will be fixed in the spec later.
To avoid going back through and repeating this process for CRP2, relevant switches were added so all of the new VUID checks could select the appropriate VUID based on whether CRP2 or the original functionality was being validated.
In doing so, it made sense to update the existing VUID checks to have the same treatment, so all checks should be covered for both the original functionality and CRP2.

Note that there is a new addition in "convert_to_renderpass2.cpp/h" which takes a VkRenderPassCreateInfo and turns it (and its substructs) into a (safe_)VkRenderPassCreateInfo2KHR automatically, allowing as little code duplication as possible such that both vkCreateRenderPass and vkCreateRenderPass2KHR both follow the same code paths.